Moved team sync from global scope.
