+++ src/engine/core/database/registry.test.ts
@@ -4,7 +4,7 @@ import { registry } from "@/engine/core/database/registry";
 
 describe("registry storage", () => {
   it("storage to contain all fields", () => {
-    expect(Object.keys(registry)).toHaveLength(45);
+    expect(Object.keys(registry)).toHaveLength(44);
   });
 
   it("storage to initialize with correct data", () => {
@@ -58,7 +58,6 @@ describe("registry storage", () => {
     expect(registry.saveMarkers).toBeInstanceOf(LuaTable);
     expect(registry.signalLights).toBeInstanceOf(LuaTable);
     expect(registry.spawnedVertexes).toBeInstanceOf(LuaTable);
-    expect(registry.patrolSynchronization).toBeInstanceOf(LuaTable);
     expect(registry.noCombatZones).toBeInstanceOf(LuaTable);
     expect(registry.noCombatSmartTerrains).toBeInstanceOf(LuaTable);
     expect(registry.baseSmartTerrains).toBeInstanceOf(LuaTable);
