+++ src/engine/core/objects/ai/patrol/PatrolConfig.ts
@@ -1,0 +1,9 @@
+import { TName, TNumberId } from "@/engine/lib/types";
+
+export const patrolConfig = {
+  /**
+   * List of synchronization groups for game patrols.
+   * Related to stalkers patrol manager.
+   */
+  PATROL_TEAMS: new LuaTable<TName, LuaTable<TNumberId, boolean>>(),
+};
