+++ src/engine/core/objects/ai/patrol/index.ts
@@ -1,2 +1,1 @@
-export * from "@/engine/core/objects/ai/state/StalkerStateManager";
 export * from "@/engine/core/objects/ai/patrol/StalkerPatrolManager";
