+++ src/engine/core/utils/patrol.ts
@@ -124,40 +124,3 @@ export function choosePatrolWaypointByFlags(
 
   return $multi(chosenPointIndex, countOfPossiblePoints);
 }
-
-/**
- * Check whether patrol team is synchronized.
- * Used to wait before signals for all patrol members to join waypoint.
- *
- * Note: dead/offline objects are excluded from sync check.
- *
- * @param teamName - optional team name to check for sync state
- * @returns whether all patrol team participants are synchronized
- */
-export function isPatrolTeamSynchronized(teamName: Optional<TName>): boolean {
-  if (!teamName) {
-    return true;
-  }
-
-  const state: Optional<LuaTable<TNumberId, boolean>> = registry.patrolSynchronization.get(teamName);
-
-  if (!state) {
-    return true;
-  }
-
-  for (const [id, isFlagged] of state) {
-    const object: Optional<GameObject> = level.object_by_id(id);
-
-    // Check sync stat of the object if it is online and alive.
-    if (object && object.alive()) {
-      if (!isFlagged) {
-        return false;
-      }
-    } else {
-      // Delete objects that cannot be synchronized.
-      registry.patrolSynchronization.get(teamName).delete(id);
-    }
-  }
-
-  return true;
-}
